# docs/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"MenuItemDto": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"nameAr": {
						"type": "string"
					},
					"category": {
						"type": "string"
					},
					"price": {
						"type": "number",
						"format": "double"
					},
					"costPrice": {
						"type": "number",
						"format": "double"
					},
					"stock": {
						"type": "number",
						"format": "double"
					},
					"lowStockThreshold": {
						"type": "number",
						"format": "double"
					},
					"trackStock": {
						"type": "boolean"
					}
				},
				"required": [
					"id",
					"name",
					"nameAr",
					"category",
					"price"
				],
				"type": "object",
				"additionalProperties": false
			},
			"MenuItemListResponse": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"data": {
						"items": {
							"$ref": "#/components/schemas/MenuItemDto"
						},
						"type": "array"
					}
				},
				"required": [
					"success",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"MenuItemResponse": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"data": {
						"$ref": "#/components/schemas/MenuItemDto"
					}
				},
				"required": [
					"success",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"MenuActionResponse": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					}
				},
				"required": [
					"success"
				],
				"type": "object",
				"additionalProperties": false
			},
			"DeductStockItemDto": {
				"properties": {
					"id": {
						"type": "string"
					},
					"qty": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"id",
					"qty"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CategoryDto": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"nameAr": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"name",
					"nameAr"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CategoryListResponse": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"data": {
						"items": {
							"$ref": "#/components/schemas/CategoryDto"
						},
						"type": "array"
					}
				},
				"required": [
					"success",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CategoryResponse": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"data": {
						"$ref": "#/components/schemas/CategoryDto"
					}
				},
				"required": [
					"success",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "playstation-system-backend",
		"version": "1.0.0",
		"contact": {}
	},
	"paths": {
...
		"/api/v1/menu/items": {
			"get": {
				"operationId": "GetItems",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MenuItemListResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Menu Items"
				],
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "SaveItem",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MenuItemResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Menu Items"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/MenuItemDto"
							}
						}
					}
				}
			}
		},
		"/api/v1/menu/items/{id}": {
			"put": {
				"operationId": "UpdateItem",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MenuItemResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Menu Items"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/MenuItemDto"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteItem",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MenuActionResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Menu Items"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/v1/menu/items/deduct-stock": {
			"post": {
				"operationId": "DeductStock",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MenuActionResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Menu Items"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"items": {
									"$ref": "#/components/schemas/DeductStockItemDto"
								},
								"type": "array"
							}
						}
					}
				}
			}
		},
		"/api/v1/menu/items/batch": {
			"post": {
				"operationId": "SaveAllItems",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MenuItemListResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Menu Items"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"items": {
									"$ref": "#/components/schemas/MenuItemDto"
								},
								"type": "array"
							}
						}
					}
				}
			}
		},
...
		"/api/v1/menu/categories": {
			"get": {
				"operationId": "GetCategories",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CategoryListResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Menu Categories"
				],
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "SaveCategory",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CategoryResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Menu Categories"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CategoryDto"
							}
						}
					}
				}
			}
		},
		"/api/v1/menu/categories/batch": {
			"post": {
				"operationId": "SaveAllCategories",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CategoryListResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Menu Categories"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"items": {
									"$ref": "#/components/schemas/CategoryDto"
								},
								"type": "array"
							}
						}
					}
				}
			}
		},
...
		"/api/v1/menu/categories/{id}": {
			"delete": {
				"operationId": "DeleteCategory",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MenuActionResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Menu Categories"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}